Refactor cd..
