+++ tests/test_models_partition.py
@@ -6,7 +6,7 @@ import sys
 
 from multiprocessing import JoinableQueue
 
-from pg_view.view import DiskCollectorConsumer
+from pg_view.consumers import DiskCollectorConsumer
 
 path = os.path.dirname(os.path.dirname(__file__))
 sys.path.insert(0, path)
